ocr-api/requirements.txt
mysqlclient
Flask
flask-sqlalchemy
sqlalchemy
google-cloud-storage
pytesseract
tesseract
pandas
numpy
opencv-python
tensorflow
python-dotenv
matplotlib
Pillow
gdown
gunicorn